fix: widen statvfs fields so the crate builds on macOS
free_bytes multiplies statvfs.f_bavail by statvfs.f_frsize, which the
comment claimed were both 64-bit on Linux and macOS. On macOS f_bavail is
u32, so both release builds failed with:

  error[E0277]: cannot multiply `u32` by `u64`
   --> src/models/downloader.rs:177:22

Widen both operands explicitly; the product is bytes on either platform.

The Linux job never caught this because f_bavail is u64 there, and the
macOS cfg branches are not compiled anywhere else. CI gains a macOS job
running `cargo check --locked --all-targets` so the macOS-only branches
are compiled on every change instead of at release time.

+++ src-tauri/src/models/downloader.rs
@@ -172,9 +172,9 @@ fn free_bytes(path: &Path) -> Result<u64, String> {
     if rc != 0 {
         return Err(format!("无法读取磁盘空间: {}", path.display()));
     }
-    // f_bavail and f_frsize are 64-bit on the Linux and macOS targets we
-    // build for, so this stays a plain u64 multiplication.
-    Ok(stat.f_bavail * stat.f_frsize)
+    // `f_bavail` is 32-bit on macOS and 64-bit on Linux; widen both sides
+    // instead of assuming they line up. The product is bytes either way.
+    Ok(u64::from(stat.f_bavail) * u64::from(stat.f_frsize))
 }
 
 /// Download from an ordered list of sources, falling back on failure.
